requirements.txt
PySide6==6.11.0
qt-material
Pillow
PyOpenGL
numpy
rawpy
imageio
imageio-ffmpeg
PyOpenGL_accelerate
Imervue